renamed files and branch as per new design
